+++ Source/AccelByteUe4Sdk/Private/Core/AccelByteEntitlementTokenGenerator.cpp
@@ -11,38 +11,60 @@ using namespace AccelByte;
 DECLARE_LOG_CATEGORY_EXTERN(LogAccelByteEntitlementTokenGenerator, Log, All);
 DEFINE_LOG_CATEGORY(LogAccelByteEntitlementTokenGenerator);
 
-FAccelByteEntitlementTokenGenerator::FAccelByteEntitlementTokenGenerator(const FTokenGeneratorParams& Params) :
-	Skus(Params.Skus),
-	ItemIds(Params.ItemIds),
-	AppIds(Params.AppIds)
-{}
-
-FAccelByteEntitlementTokenGenerator::~FAccelByteEntitlementTokenGenerator()
+FAccelByteEntitlementTokenGenerator::FAccelByteEntitlementTokenGenerator(const FTokenGeneratorParams& Params)
+	: FAccelByteEntitlementTokenGenerator(Params, nullptr)
 {
-	IAccelByteTokenGenerator::~IAccelByteTokenGenerator();
+}
+
+FAccelByteEntitlementTokenGenerator::FAccelByteEntitlementTokenGenerator(const FTokenGeneratorParams& Params, const FApiClientPtr& InApiClient)
+	: Skus(Params.Skus)
+	, ItemIds(Params.ItemIds)
+	, AppIds(Params.AppIds)
+	, ApiClient(InApiClient)
+{
 }
 
 void FAccelByteEntitlementTokenGenerator::RequestToken()
 {
-	if(!Token.IsEmpty() && ExpiredDateTime > FDateTime::UtcNow())
+	if (!Token.IsEmpty() && ExpiredDateTime > FDateTime::UtcNow())
 	{
 		TokenReceivedEvent.Broadcast(Token);
 		return;
 	}
 
-	FRegistry::Entitlement.GetUserEntitlementOwnershipTokenOnly(ItemIds, AppIds, Skus,
-		THandler<FAccelByteModelsOwnershipToken>::CreateLambda([&](const FAccelByteModelsOwnershipToken& result)
-		{
-			Token = result.OwnershipToken;
-			FDateTime TokenExpiration;
-			AccelByteJwtWrapper::GetExpirationDateTime(Token, TokenExpiration);
+	const THandler<FAccelByteModelsOwnershipToken> OnGetUserEntitlementOwnershipTokenSuccessDelegate =
+		THandler<FAccelByteModelsOwnershipToken>::CreateSP(SharedThis(this), &FAccelByteEntitlementTokenGenerator::OnGetUserEntitlementOwnershipTokenSuccess);
 
-			ExpiredDateTime = FDateTime::UtcNow() + (TokenExpiration - FDateTime::UtcNow() - FTimespan::FromSeconds(30));
-			TokenReceivedEvent.Broadcast(Token);
-		}), FErrorHandler::CreateLambda([](const int32 Code, const FString& Message)
-		{
-			UE_LOG(LogAccelByteEntitlementTokenGenerator, Warning, TEXT("Get user entitlement ownership token failed with code %d, %s"), Code, *Message);
-		}));
+	const FErrorHandler OnGetUserEntitlementOwnershipTokenErrorDelegate =
+		FErrorHandler::CreateSP(SharedThis(this), &FAccelByteEntitlementTokenGenerator::OnGetUserEntitlementOwnershipTokenError);
+
+	if (ApiClient.IsValid())
+	{
+		ApiClient->Entitlement.GetUserEntitlementOwnershipTokenOnly(ItemIds
+			, AppIds
+			, Skus
+			, OnGetUserEntitlementOwnershipTokenSuccessDelegate
+			, OnGetUserEntitlementOwnershipTokenErrorDelegate);
+	}
+	else
+	{
+		FRegistry::Entitlement.GetUserEntitlementOwnershipTokenOnly(ItemIds
+			, AppIds
+			, Skus
+			, OnGetUserEntitlementOwnershipTokenSuccessDelegate
+			, OnGetUserEntitlementOwnershipTokenErrorDelegate);
+	}
+}
+
+void FAccelByteEntitlementTokenGenerator::OnGetUserEntitlementOwnershipTokenSuccess(const FAccelByteModelsOwnershipToken& Result)
+{
+	Token = Result.OwnershipToken;
+
+	FDateTime TokenExpiration{};
+	AccelByteJwtWrapper::GetExpirationDateTime(Token, TokenExpiration);
+	ExpiredDateTime = FDateTime::UtcNow() + (TokenExpiration - FDateTime::UtcNow() - FTimespan::FromSeconds(30));
+
+	TokenReceivedEvent.Broadcast(Token);
 }
 
 bool FAccelByteEntitlementTokenGenerator::IsTokenValid() const
@@ -53,4 +75,13 @@ bool FAccelByteEntitlementTokenGenerator::IsTokenValid() const
 FAccelByteEntitlementTokenGenerator::FTokenReceivedEvent& FAccelByteEntitlementTokenGenerator::OnTokenReceived()
 {
 	return TokenReceivedEvent;
+}
+
+void FAccelByteEntitlementTokenGenerator::OnGetUserEntitlementOwnershipTokenError(int32 ErrorCode, const FString& ErrorMessage)
+{
+	UE_LOG(LogAccelByteEntitlementTokenGenerator, Warning, TEXT("Get user entitlement ownership token failed with code %d, %s"), ErrorCode, *ErrorMessage);
+
+	// Even though token generation failed, we want to broadcast our 'TokenReceivedEvent' with an invalid token string.
+	// This way, lobby can properly raise a connection error for the failed attempt.
+	TokenReceivedEvent.Broadcast(ErrorTokenRequestFailed);
 }